@@ -12,7 +12,6 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/segmentio/asm/base64"
@@ -39,7 +38,7 @@ type guildMemberPageToken struct {
 	nextQueryParam string
 }
 
-var defaultPageToken = &guildMemberPageToken{
+var defaultPageToken = guildMemberPageToken{
 	guildID:    0,
 	nextIndex:  0,
 	nextRoleID: 2,
@@ -151,13 +150,13 @@ func (bot *CQBot) CQGetGuildMembers(guildID uint64, nextToken string) global.MSG
 	if guild == nil {
 		return Failed(100, "GUILD_NOT_FOUND")
 	}
-	token := defaultPageToken
+	token := &defaultPageToken
 	if nextToken != "" {
 		i, exists := bot.nextTokenCache.Get(nextToken)
 		if !exists {
 			return Failed(100, "NEXT_TOKEN_NOT_EXISTS")
 		}
-		token = i.(*guildMemberPageToken)
+		token = i
 		if token.guildID != guildID {
 			return Failed(100, "GUILD_NOT_MATCH")
 		}
@@ -408,7 +407,6 @@ func (bot *CQBot) CQGetGroupList(noCache bool) global.MSG {
 		gs = append(gs, global.MSG{
 			"group_id":          g.Code,
 			"group_name":        g.Name,
-			"group_memo":        g.Memo,
 			"group_create_time": g.GroupCreateTime,
 			"group_level":       g.GroupLevel,
 			"max_member_count":  g.MaxMemberCount,
@@ -450,7 +448,6 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) global.MSG {
 		return OK(global.MSG{
 			"group_id":          group.Code,
 			"group_name":        group.Name,
-			"group_memo":        group.Memo,
 			"group_create_time": group.GroupCreateTime,
 			"group_level":       group.GroupLevel,
 			"max_member_count":  group.MaxMemberCount,
@@ -687,6 +684,24 @@ func (bot *CQBot) CQSendMessage(groupID, userID int64, m gjson.Result, messageTy
 	return global.MSG{}
 }
 
+// CQSendForwardMessage 发送合并转发消息
+//
+// @route(send_forward_msg)
+// @rename(m->messages)
+func (bot *CQBot) CQSendForwardMessage(groupID, userID int64, m gjson.Result, messageType string) global.MSG {
+	switch {
+	case messageType == "group":
+		return bot.CQSendGroupForwardMessage(groupID, m)
+	case messageType == "private":
+		fallthrough
+	case userID != 0:
+		return bot.CQSendPrivateForwardMessage(userID, m)
+	case groupID != 0:
+		return bot.CQSendGroupForwardMessage(groupID, m)
+	}
+	return global.MSG{}
+}
+
 // CQSendGroupMessage 发送群消息
 //
 // https://git.io/Jtz1c
@@ -712,7 +727,7 @@ func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape b
 
 	var elem []message.IMessageElement
 	if m.Type == gjson.JSON {
-		elem = bot.ConvertObjectMessage(m, MessageSourceGroup)
+		elem = bot.ConvertObjectMessage(m, message.SourceGroup)
 	} else {
 		str := m.String()
 		if str == "" {
@@ -722,7 +737,7 @@ func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape b
 		if autoEscape {
 			elem = []message.IMessageElement{message.NewText(str)}
 		} else {
-			elem = bot.ConvertStringMessage(str, MessageSourceGroup)
+			elem = bot.ConvertStringMessage(str, message.SourceGroup)
 		}
 	}
 	fixAt(elem)
@@ -766,7 +781,7 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 
 	var elem []message.IMessageElement
 	if m.Type == gjson.JSON {
-		elem = bot.ConvertObjectMessage(m, MessageSourceGuildChannel)
+		elem = bot.ConvertObjectMessage(m, message.SourceGuildChannel)
 	} else {
 		str := m.String()
 		if str == "" {
@@ -776,7 +791,7 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 		if autoEscape {
 			elem = []message.IMessageElement{message.NewText(str)}
 		} else {
-			elem = bot.ConvertStringMessage(str, MessageSourceGuildChannel)
+			elem = bot.ConvertStringMessage(str, message.SourceGuildChannel)
 		}
 	}
 	fixAt(elem)
@@ -788,116 +803,126 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 	return OK(global.MSG{"message_id": mid})
 }
 
-func (bot *CQBot) uploadForwardElement(m gjson.Result, groupID int64) *message.ForwardElement {
+func (bot *CQBot) uploadForwardElement(m gjson.Result, target int64, sourceType message.SourceType) *message.ForwardElement {
 	ts := time.Now().Add(-time.Minute * 5)
-	fm := message.NewForwardMessage()
-
-	var lazyUpload []func()
-	var wg sync.WaitGroup
-	resolveElement := func(elems []message.IMessageElement) []message.IMessageElement {
-		for i, elem := range elems {
-			iescape := i
-			switch o := elem.(type) {
-			case *LocalImageElement, *LocalVideoElement:
-				wg.Add(1)
-				lazyUpload = append(lazyUpload, func() {
-					defer wg.Done()
-					gm, err := bot.uploadMedia(o, groupID, true)
-					if err != nil {
-						log.Warnf("警告: 群 %d %s上传失败: %v", groupID, o.Type().String(), err)
-					} else {
-						elems[iescape] = gm
-					}
-				})
-			}
-		}
-		return elems
+	groupID := target
+	source := message.Source{SourceType: sourceType, PrimaryID: target}
+	if sourceType == message.SourcePrivate {
+		groupID = 0
 	}
+	builder := bot.Client.NewForwardMessageBuilder(groupID)
 
-	convert := func(e gjson.Result) *message.ForwardNode {
-		if e.Get("type").Str != "node" {
-			return nil
-		}
-		ts.Add(time.Second)
-		if e.Get("data.id").Exists() {
-			i := e.Get("data.id").Int()
-			m, _ := db.GetGroupMessageByGlobalID(int32(i))
-			if m != nil {
-				return &message.ForwardNode{
-					SenderId:   m.Attribute.SenderUin,
-					SenderName: m.Attribute.SenderName,
-					Time: func() int32 {
-						msgTime := m.Attribute.Timestamp
-						if msgTime == 0 {
-							return int32(ts.Unix())
+	var convertMessage func(m gjson.Result) *message.ForwardMessage
+	convertMessage = func(m gjson.Result) *message.ForwardMessage {
+		fm := message.NewForwardMessage()
+		var w worker
+		resolveElement := func(elems []message.IMessageElement) []message.IMessageElement {
+			for i, elem := range elems {
+				p := &elems[i]
+				switch o := elem.(type) {
+				case *LocalVideoElement:
+					w.do(func() {
+						gm, err := bot.uploadLocalVideo(source, o)
+						if err != nil {
+							log.Warnf(uploadFailedTemplate, "合并转发", target, "视频", err)
+						} else {
+							*p = gm
 						}
-						return int32(msgTime)
-					}(),
-					Message: resolveElement(bot.ConvertContentMessage(m.Content, MessageSourceGroup)),
+					})
+				case *LocalImageElement:
+					w.do(func() {
+						gm, err := bot.uploadLocalImage(source, o)
+						if err != nil {
+							log.Warnf(uploadFailedTemplate, "合并转发", target, "图片", err)
+						} else {
+							*p = gm
+						}
+					})
 				}
 			}
-			log.Warnf("警告: 引用消息 %v 错误或数据库未开启.", e.Get("data.id").Str)
-			return nil
+			return elems
 		}
-		uin := e.Get("data.[user_id,uin].0").Int()
-		msgTime := e.Get("data.time").Int()
-		if msgTime == 0 {
-			msgTime = ts.Unix()
-		}
-		name := e.Get("data.name").Str
-		c := e.Get("data.content")
-		if c.IsArray() {
-			nested := false
-			c.ForEach(func(_, value gjson.Result) bool {
-				if value.Get("type").Str == "node" {
-					nested = true
-					return false
+
+		convert := func(e gjson.Result) *message.ForwardNode {
+			if e.Get("type").Str != "node" {
+				return nil
+			}
+			if e.Get("data.id").Exists() {
+				i := e.Get("data.id").Int()
+				m, _ := db.GetGroupMessageByGlobalID(int32(i))
+				if m != nil {
+					msgTime := m.Attribute.Timestamp
+					if msgTime == 0 {
+						msgTime = ts.Unix()
+					}
+					return &message.ForwardNode{
+						SenderId:   m.Attribute.SenderUin,
+						SenderName: m.Attribute.SenderName,
+						Time:       int32(msgTime),
+						Message:    resolveElement(bot.ConvertContentMessage(m.Content, message.SourceGroup)),
+					}
 				}
-				return true
-			})
-			if nested { // 处理嵌套
-				fe := bot.uploadForwardElement(c, groupID)
+				log.Warnf("警告: 引用消息 %v 错误或数据库未开启.", e.Get("data.id").Str)
+				return nil
+			}
+			uin := e.Get("data.[user_id,uin].0").Int()
+			msgTime := e.Get("data.time").Int()
+			if msgTime == 0 {
+				msgTime = ts.Unix()
+			}
+			name := e.Get("data.[name,nickname].0").Str
+			c := e.Get("data.content")
+			if c.IsArray() {
+				nested := false
+				c.ForEach(func(_, value gjson.Result) bool {
+					if value.Get("type").Str == "node" {
+						nested = true
+						return false
+					}
+					return true
+				})
+				if nested { // 处理嵌套
+					nestedNode := builder.NestedNode()
+					builder.Link(nestedNode, convertMessage(c))
+					return &message.ForwardNode{
+						SenderId:   uin,
+						SenderName: name,
+						Time:       int32(msgTime),
+						Message:    []message.IMessageElement{nestedNode},
+					}
+				}
+			}
+			content := bot.ConvertObjectMessage(c, message.SourceGroup)
+			if uin != 0 && name != "" && len(content) > 0 {
 				return &message.ForwardNode{
 					SenderId:   uin,
 					SenderName: name,
 					Time:       int32(msgTime),
-					Message:    []message.IMessageElement{fe},
+					Message:    resolveElement(content),
 				}
 			}
+			log.Warnf("警告: 非法 Forward node 将跳过. uin: %v name: %v content count: %v", uin, name, len(content))
+			return nil
 		}
-		content := bot.ConvertObjectMessage(c, MessageSourceGroup)
-		if uin != 0 && name != "" && len(content) > 0 {
-			return &message.ForwardNode{
-				SenderId:   uin,
-				SenderName: name,
-				Time:       int32(msgTime),
-				Message:    resolveElement(content),
-			}
-		}
-		log.Warnf("警告: 非法 Forward node 将跳过. uin: %v name: %v content count: %v", uin, name, len(content))
-		return nil
-	}
 
-	if m.IsArray() {
-		for _, item := range m.Array() {
-			node := convert(item)
+		if m.IsArray() {
+			for _, item := range m.Array() {
+				node := convert(item)
+				if node != nil {
+					fm.AddNode(node)
+				}
+			}
+		} else {
+			node := convert(m)
 			if node != nil {
 				fm.AddNode(node)
 			}
 		}
-	} else {
-		node := convert(m)
-		if node != nil {
-			fm.AddNode(node)
-		}
-	}
 
-	for _, upload := range lazyUpload {
-		go upload()
+		w.wait()
+		return fm
 	}
-	wg.Wait()
-
-	return bot.Client.UploadGroupForwardMessage(groupID, fm)
+	return builder.Main(convertMessage(m))
 }
 
 // CQSendGroupForwardMessage 扩展API-发送合并转发(群)
@@ -910,18 +935,39 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) globa
 		return Failed(100)
 	}
 
-	fe := bot.uploadForwardElement(m, groupID)
-	if fe != nil {
-		ret := bot.Client.SendGroupForwardMessage(groupID, fe)
-		if ret == nil || ret.Id == -1 {
-			log.Warnf("合并转发(群)消息发送失败: 账号可能被风控.")
-			return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
-		}
-		return OK(global.MSG{
-			"message_id": bot.InsertGroupMessage(ret),
-		})
+	fe := bot.uploadForwardElement(m, groupID, message.SourceGroup)
+	if fe == nil {
+		return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
 	}
-	return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
+	ret := bot.Client.SendGroupForwardMessage(groupID, fe)
+	if ret == nil || ret.Id == -1 {
+		log.Warnf("合并转发(群)消息发送失败: 账号可能被风控.")
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
+	}
+	return OK(global.MSG{
+		"message_id": bot.InsertGroupMessage(ret),
+	})
+}
+
+// CQSendPrivateForwardMessage 扩展API-发送合并转发(好友)
+//
+// https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E5%90%88%E5%B9%B6%E8%BD%AC%E5%8F%91-%E7%BE%A4
+// @route(send_private_forward_msg)
+// @rename(m->messages)
+func (bot *CQBot) CQSendPrivateForwardMessage(userID int64, m gjson.Result) global.MSG {
+	if m.Type != gjson.JSON {
+		return Failed(100)
+	}
+	fe := bot.uploadForwardElement(m, userID, message.SourcePrivate)
+	if fe == nil {
+		return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
+	}
+	mid := bot.SendPrivateMessage(userID, 0, &message.SendingMessage{Elements: []message.IMessageElement{fe}})
+	if mid == -1 {
+		log.Warnf("合并转发(好友)消息发送失败: 账号可能被风控.")
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
+	}
+	return OK(global.MSG{"message_id": mid})
 }
 
 // CQSendPrivateMessage 发送私聊消息
@@ -932,7 +978,7 @@ func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) globa
 func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	var elem []message.IMessageElement
 	if m.Type == gjson.JSON {
-		elem = bot.ConvertObjectMessage(m, MessageSourcePrivate)
+		elem = bot.ConvertObjectMessage(m, message.SourcePrivate)
 	} else {
 		str := m.String()
 		if str == "" {
@@ -941,7 +987,7 @@ func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Resu
 		if autoEscape {
 			elem = []message.IMessageElement{message.NewText(str)}
 		} else {
-			elem = bot.ConvertStringMessage(str, MessageSourcePrivate)
+			elem = bot.ConvertStringMessage(str, message.SourcePrivate)
 		}
 	}
 	mid := bot.SendPrivateMessage(userID, groupID, &message.SendingMessage{Elements: elem})
@@ -992,6 +1038,17 @@ func (bot *CQBot) CQSetGroupName(groupID int64, name string) global.MSG {
 		return OK(nil)
 	}
 	return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
+}
+
+// CQGetGroupMemo 扩展API-获取群公告
+// @route(_get_group_notice)
+func (bot *CQBot) CQGetGroupMemo(groupID int64) global.MSG {
+	r, err := bot.Client.GetGroupNotice(groupID)
+	if err != nil {
+		return Failed(100, "获取群公告失败", err.Error())
+	}
+
+	return OK(r)
 }
 
 // CQSetGroupMemo 扩展API-发送群公告
@@ -1114,9 +1171,9 @@ func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) global.MSG {
 		return Failed(100, "FLAG_NOT_FOUND", "FLAG不存在")
 	}
 	if approve {
-		req.(*client.NewFriendRequest).Accept()
+		req.Accept()
 	} else {
-		req.(*client.NewFriendRequest).Reject()
+		req.Reject()
 	}
 	return OK(nil)
 }
@@ -1223,27 +1280,6 @@ func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) global.MSG
 	return OK(nil)
 }
 
-// CQGetVipInfo 扩展API-获取VIP信息
-//
-// https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96vip%E4%BF%A1%E6%81%AF
-// @route(_get_vip_info)
-func (bot *CQBot) CQGetVipInfo(userID int64) global.MSG {
-	vip, err := bot.Client.GetVipInfo(userID)
-	if err != nil {
-		return Failed(100, "VIP_API_ERROR", err.Error())
-	}
-	msg := global.MSG{
-		"user_id":          vip.Uin,
-		"nickname":         vip.Name,
-		"level":            vip.Level,
-		"level_speed":      vip.LevelSpeed,
-		"vip_level":        vip.VipLevel,
-		"vip_growth_speed": vip.VipGrowthSpeed,
-		"vip_growth_total": vip.VipGrowthTotal,
-	}
-	return OK(msg)
-}
-
 // CQGetGroupHonorInfo 获取群荣誉信息
 //
 // https://git.io/Jtz1H
@@ -1347,7 +1383,7 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global
 
 		if reply.Exists() {
 			autoEscape := param.EnsureBool(operation.Get("auto_escape"), false)
-			at := operation.Get("at_sender").Bool() && !isAnonymous && msgType == "group"
+			at := !isAnonymous && operation.Get("at_sender").Bool() && msgType == "group"
 			if at && reply.IsArray() {
 				// 在 reply 数组头部插入CQ码
 				replySegments := make([]global.MSG, 0)
@@ -1393,7 +1429,7 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global
 			if operation.Get("delete").Bool() {
 				bot.CQDeleteMessage(int32(context.Get("message_id").Int()))
 			}
-			if operation.Get("kick").Bool() && !isAnonymous {
+			if !isAnonymous && operation.Get("kick").Bool() {
 				bot.CQSetGroupKick(context.Get("group_id").Int(), context.Get("user_id").Int(), "", operation.Get("reject_add_request").Bool())
 			}
 			if operation.Get("ban").Bool() {
@@ -1475,18 +1511,18 @@ func (bot *CQBot) CQDownloadFile(url string, headers gjson.Result, threadCount i
 	h := map[string]string{}
 	if headers.IsArray() {
 		for _, sub := range headers.Array() {
-			str := strings.SplitN(sub.String(), "=", 2)
-			if len(str) == 2 {
-				h[str[0]] = str[1]
+			first, second, ok := strings.Cut(sub.String(), "=")
+			if ok {
+				h[first] = second
 			}
 		}
 	}
 	if headers.Type == gjson.String {
 		lines := strings.Split(headers.String(), "\r\n")
 		for _, sub := range lines {
-			str := strings.SplitN(sub, "=", 2)
-			if len(str) == 2 {
-				h[str[0]] = str[1]
+			first, second, ok := strings.Cut(sub, "=")
+			if ok {
+				h[first] = second
 			}
 		}
 	}
@@ -1525,7 +1561,7 @@ func (bot *CQBot) CQGetForwardMessage(resID string) global.MSG {
 		r := make([]global.MSG, len(nodes))
 		for i, n := range nodes {
 			bot.checkMedia(n.Message, 0)
-			content := ToFormattedMessage(n.Message, MessageSource{SourceType: MessageSourceGroup}, false)
+			content := ToFormattedMessage(n.Message, message.Source{SourceType: message.SourceGroup})
 			if len(n.Message) == 1 {
 				if forward, ok := n.Message[0].(*message.ForwardMessage); ok {
 					content = transformNodes(forward.Nodes)
@@ -1536,8 +1572,9 @@ func (bot *CQBot) CQGetForwardMessage(resID string) global.MSG {
 					"user_id":  n.SenderId,
 					"nickname": n.SenderName,
 				},
-				"time":    n.Time,
-				"content": content,
+				"time":     n.Time,
+				"content":  content,
+				"group_id": n.GroupId,
 			}
 		}
 		return r
@@ -1573,9 +1610,9 @@ func (bot *CQBot) CQGetMessage(messageID int32) global.MSG {
 	switch o := msg.(type) {
 	case *db.StoredGroupMessage:
 		m["group_id"] = o.GroupCode
-		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourceGroup), MessageSource{SourceType: MessageSourceGroup, PrimaryID: uint64(o.GroupCode)}, false)
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, message.SourceGroup), message.Source{SourceType: message.SourceGroup, PrimaryID: o.GroupCode})
 	case *db.StoredPrivateMessage:
-		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourcePrivate), MessageSource{SourceType: MessageSourcePrivate}, false)
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, message.SourcePrivate), message.Source{SourceType: message.SourcePrivate})
 	}
 	return OK(m)
 }
@@ -1584,28 +1621,28 @@ func (bot *CQBot) CQGetMessage(messageID int32) global.MSG {
 // @route(get_guild_msg)
 func (bot *CQBot) CQGetGuildMessage(messageID string, noCache bool) global.MSG {
 	source, seq := decodeGuildMessageID(messageID)
-	if source == nil {
+	if source.SourceType == 0 {
 		log.Warnf("获取消息时出现错误: 无效消息ID")
 		return Failed(100, "INVALID_MESSAGE_ID", "无效消息ID")
 	}
 	m := global.MSG{
 		"message_id": messageID,
 		"message_source": func() string {
-			if source.SourceType == MessageSourceGuildDirect {
+			if source.SourceType == message.SourceGuildDirect {
 				return "direct"
 			}
 			return "channel"
 		}(),
 		"message_seq": seq,
-		"guild_id":    fU64(source.PrimaryID),
+		"guild_id":    fU64(uint64(source.PrimaryID)),
 		"reactions":   []int{},
 	}
 	// nolint: exhaustive
 	switch source.SourceType {
-	case MessageSourceGuildChannel:
-		m["channel_id"] = fU64(source.SubID)
+	case message.SourceGuildChannel:
+		m["channel_id"] = fU64(uint64(source.SecondaryID))
 		if noCache {
-			pull, err := bot.Client.GuildService.PullGuildChannelMessage(source.PrimaryID, source.SubID, seq, seq)
+			pull, err := bot.Client.GuildService.PullGuildChannelMessage(uint64(source.PrimaryID), uint64(source.SecondaryID), seq, seq)
 			if err != nil {
 				log.Warnf("获取消息时出现错误: %v", err)
 				return Failed(100, "API_ERROR", err.Error())
@@ -1620,7 +1657,7 @@ func (bot *CQBot) CQGetGuildMessage(messageID string, noCache bool) global.MSG {
 				"tiny_id":  fU64(pull[0].Sender.TinyId),
 				"nickname": pull[0].Sender.Nickname,
 			}
-			m["message"] = ToFormattedMessage(pull[0].Elements, *source, false)
+			m["message"] = ToFormattedMessage(pull[0].Elements, source)
 			m["reactions"] = convertReactions(pull[0].Reactions)
 			bot.InsertGuildChannelMessage(pull[0])
 		} else {
@@ -1635,11 +1672,11 @@ func (bot *CQBot) CQGetGuildMessage(messageID string, noCache bool) global.MSG {
 				"tiny_id":  fU64(channelMsgByDB.Attribute.SenderTinyID),
 				"nickname": channelMsgByDB.Attribute.SenderName,
 			}
-			m["message"] = ToFormattedMessage(bot.ConvertContentMessage(channelMsgByDB.Content, MessageSourceGuildChannel), *source)
+			m["message"] = ToFormattedMessage(bot.ConvertContentMessage(channelMsgByDB.Content, message.SourceGuildChannel), source)
 		}
-	case MessageSourceGuildDirect:
+	case message.SourceGuildDirect:
 		// todo(mrs4s): 支持 direct 消息
-		m["tiny_id"] = fU64(source.SubID)
+		m["tiny_id"] = fU64(uint64(source.SecondaryID))
 	}
 	return OK(m)
 }
@@ -1678,12 +1715,12 @@ func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) global.MSG 
 		log.Warnf("获取群历史消息失败: %v", err)
 		return Failed(100, "MESSAGES_API_ERROR", err.Error())
 	}
-	ms := make([]global.MSG, 0, len(msg))
+	ms := make([]*event, 0, len(msg))
 	for _, m := range msg {
 		bot.checkMedia(m.Elements, groupID)
 		id := bot.InsertGroupMessage(m)
 		t := bot.formatGroupMessage(m)
-		t["message_id"] = id
+		t.Others["message_id"] = id
 		ms = append(ms, t)
 	}
 	return OK(global.MSG{
@@ -1737,7 +1774,7 @@ func (bot *CQBot) CQCanSendRecord() global.MSG {
 // @route(ocr_image,".ocr_image")
 // @rename(image_id->image)
 func (bot *CQBot) CQOcrImage(imageID string) global.MSG {
-	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, MessageSourceGroup)
+	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, message.SourceGroup)
 	if err != nil {
 		log.Warnf("load image error: %v", err)
 		return Failed(100, "LOAD_FILE_ERROR", err.Error())
@@ -1777,12 +1814,10 @@ func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration in
 		return Failed(100, "INVALID_FLAG", "无效的flag")
 	}
 	if g := bot.Client.FindGroup(groupID); g != nil {
-		s := strings.SplitN(flag, "|", 2)
-		if len(s) != 2 {
+		id, nick, ok := strings.Cut(flag, "|")
+		if !ok {
 			return Failed(100, "INVALID_FLAG", "无效的flag")
 		}
-		id := s[0]
-		nick := s[1]
 		if err := g.MuteAnonymous(id, nick, duration); err != nil {
 			log.Warnf("anonymous ban error: %v", err)
 			return Failed(100, "CALL_API_ERROR", err.Error())
@@ -1939,6 +1974,15 @@ func (bot *CQBot) CQGetModelShow(model string) global.MSG {
 	})
 }
 
+// CQSendGroupSign 群打卡
+//
+// https://club.vip.qq.com/onlinestatus/set
+// @route(send_group_sign)
+func (bot *CQBot) CQSendGroupSign(groupID int64) global.MSG {
+	bot.Client.SendGroupSign(groupID)
+	return OK(nil)
+}
+
 // CQSetModelShow 设置在线机型
 //
 // https://club.vip.qq.com/onlinestatus/set
@@ -1969,6 +2013,27 @@ func (bot *CQBot) CQMarkMessageAsRead(msgID int32) global.MSG {
 	return OK(nil)
 }
 
+// CQSetQQProfile 设置 QQ 资料
+//
+// @route(set_qq_profile)
+func (bot *CQBot) CQSetQQProfile(nickname, company, email, college, personalNote gjson.Result) global.MSG {
+	u := client.NewProfileDetailUpdate()
+
+	fi := func(f gjson.Result, do func(value string) client.ProfileDetailUpdate) {
+		if f.Exists() {
+			do(f.String())
+		}
+	}
+
+	fi(nickname, u.Nick)
+	fi(company, u.Company)
+	fi(email, u.Email)
+	fi(college, u.College)
+	fi(personalNote, u.PersonalNote)
+	bot.Client.UpdateProfile(u)
+	return OK(nil)
+}
+
 // CQReloadEventFilter 重载事件过滤器
 //
 // @route(reload_event_filter)
@@ -1984,8 +2049,7 @@ func OK(data interface{}) global.MSG {
 
 // Failed 生成失败返回值
 func Failed(code int, msg ...string) global.MSG {
-	m := ""
-	w := ""
+	m, w := "", ""
 	if len(msg) > 0 {
 		m = msg[0]
 	}
