@@ -5,15 +5,18 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/Mrs4s/MiraiGo/message"
 	"github.com/Mrs4s/MiraiGo/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/tidwall/gjson"
+
+	"github.com/Mrs4s/go-cqhttp/coolq/cqcode"
 )
 
 var bot = CQBot{}
 
 func TestCQBot_ConvertStringMessage(t *testing.T) {
-	for _, v := range bot.ConvertStringMessage(`[CQ:face,id=115,text=111][CQ:face,id=217]] [CQ:text,text=123] [`, MessageSourcePrivate) {
+	for _, v := range bot.ConvertStringMessage(`[CQ:face,id=115,text=111][CQ:face,id=217]] [CQ:text,text=123] [`, message.SourcePrivate) {
 		fmt.Println(v)
 	}
 }
@@ -25,14 +28,14 @@ var (
 
 func BenchmarkCQBot_ConvertStringMessage(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		bot.ConvertStringMessage(bench, MessageSourcePrivate)
+		bot.ConvertStringMessage(bench, message.SourcePrivate)
 	}
 	b.SetBytes(int64(len(bench)))
 }
 
 func BenchmarkCQBot_ConvertObjectMessage(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		bot.ConvertObjectMessage(benchArray, MessageSourcePrivate)
+		bot.ConvertObjectMessage(benchArray, message.SourcePrivate)
 	}
 }
 
@@ -41,7 +44,7 @@ const bText = `123456789[]&987654321[]&987654321[]&987654321[]&987654321[]&98765
 func BenchmarkCQCodeEscapeText(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		ret := bText
-		CQCodeEscapeText(ret)
+		cqcode.EscapeText(ret)
 	}
 }
 
@@ -61,6 +64,6 @@ func TestCQCodeEscapeText(t *testing.T) {
 		ret = strings.ReplaceAll(ret, "&", "&amp;")
 		ret = strings.ReplaceAll(ret, "[", "&#91;")
 		ret = strings.ReplaceAll(ret, "]", "&#93;")
-		assert.Equal(t, ret, CQCodeEscapeText(rs))
+		assert.Equal(t, ret, cqcode.EscapeText(rs))
 	}
 }
