@@ -2,12 +2,12 @@ package coolq
 
 import (
 	"encoding/hex"
+	"encoding/json"
 	"fmt"
 	"os"
 	"path"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/Mrs4s/MiraiGo/binary"
 	"github.com/Mrs4s/MiraiGo/client"
@@ -21,41 +21,73 @@ import (
 )
 
 // ToFormattedMessage 将给定[]message.IMessageElement转换为通过coolq.SetMessageFormat所定义的消息上报格式
-func ToFormattedMessage(e []message.IMessageElement, source MessageSource, isRaw ...bool) (r interface{}) {
+func ToFormattedMessage(e []message.IMessageElement, source message.Source) (r interface{}) {
 	if base.PostFormat == "string" {
-		r = ToStringMessage(e, source, isRaw...)
+		r = toStringMessage(e, source)
 	} else if base.PostFormat == "array" {
-		r = ToArrayMessage(e, source)
+		r = toElements(e, source)
 	}
 	return
 }
 
-func (bot *CQBot) privateMessageEvent(c *client.QQClient, m *message.PrivateMessage) {
-	bot.checkMedia(m.Elements, m.Sender.Uin)
-	source := MessageSource{
-		SourceType: MessageSourcePrivate,
-		PrimaryID:  uint64(m.Sender.Uin),
+type event struct {
+	PostType   string
+	DetailType string
+	SubType    string
+	Time       int64
+	SelfID     int64
+	Others     global.MSG
+}
+
+func (ev *event) MarshalJSON() ([]byte, error) {
+	buf := global.NewBuffer()
+	defer global.PutBuffer(buf)
+
+	detail := ""
+	switch ev.PostType {
+	case "message", "message_sent":
+		detail = "message_type"
+	case "notice":
+		detail = "notice_type"
+	case "request":
+		detail = "request_type"
+	case "meta_event":
+		detail = "meta_event_type"
+	default:
+		panic("unknown post type: " + ev.PostType)
 	}
-	cqm := ToStringMessage(m.Elements, source, true)
+	fmt.Fprintf(buf, `{"post_type":"%s","%s":"%s","time":%d,"self_id":%d`, ev.PostType, detail, ev.DetailType, ev.Time, ev.SelfID)
+	if ev.SubType != "" {
+		fmt.Fprintf(buf, `,"sub_type":"%s"`, ev.SubType)
+	}
+	for k, v := range ev.Others {
+		v, _ := json.Marshal(v)
+		fmt.Fprintf(buf, `,"%s":%s`, k, v)
+	}
+	buf.WriteByte('}')
+	return append([]byte(nil), buf.Bytes()...), nil
+}
+
+func (bot *CQBot) privateMessageEvent(_ *client.QQClient, m *message.PrivateMessage) {
+	bot.checkMedia(m.Elements, m.Sender.Uin)
+	source := message.Source{
+		SourceType: message.SourcePrivate,
+		PrimaryID:  m.Sender.Uin,
+	}
+	cqm := toStringMessage(m.Elements, source)
 	id := bot.InsertPrivateMessage(m)
 	log.Infof("收到好友 %v(%v) 的消息: %v (%v)", m.Sender.DisplayName(), m.Sender.Uin, cqm, id)
+	typ := "message/private/friend"
+	if m.Sender.Uin == bot.Client.Uin {
+		typ = "message_sent/private/friend"
+	}
 	fm := global.MSG{
-		"post_type": func() string {
-			if m.Sender.Uin == bot.Client.Uin {
-				return "message_sent"
-			}
-			return "message"
-		}(),
-		"message_type": "private",
-		"sub_type":     "friend",
-		"message_id":   id,
-		"user_id":      m.Sender.Uin,
-		"target_id":    m.Target,
-		"message":      ToFormattedMessage(m.Elements, source, false),
-		"raw_message":  cqm,
-		"font":         0,
-		"self_id":      c.Uin,
-		"time":         time.Now().Unix(),
+		"message_id":  id,
+		"user_id":     m.Sender.Uin,
+		"target_id":   m.Target,
+		"message":     ToFormattedMessage(m.Elements, source),
+		"raw_message": cqm,
+		"font":        0,
 		"sender": global.MSG{
 			"user_id":  m.Sender.Uin,
 			"nickname": m.Sender.Nickname,
@@ -63,7 +95,7 @@ func (bot *CQBot) privateMessageEvent(c *client.QQClient, m *message.PrivateMess
 			"age":      0,
 		},
 	}
-	bot.dispatchEventMessage(fm)
+	bot.dispatchEvent(typ, fm)
 }
 
 func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage) {
@@ -71,11 +103,9 @@ func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage)
 	for _, elem := range m.Elements {
 		if file, ok := elem.(*message.GroupFileElement); ok {
 			log.Infof("群 %v(%v) 内 %v(%v) 上传了文件: %v", m.GroupName, m.GroupCode, m.Sender.DisplayName(), m.Sender.Uin, file.Name)
-			bot.dispatchEventMessage(global.MSG{
-				"post_type":   "notice",
-				"notice_type": "group_upload",
-				"group_id":    m.GroupCode,
-				"user_id":     m.Sender.Uin,
+			bot.dispatchEvent("notice/group_upload", global.MSG{
+				"group_id": m.GroupCode,
+				"user_id":  m.Sender.Uin,
 				"file": global.MSG{
 					"id":    file.Path,
 					"name":  file.Name,
@@ -83,35 +113,33 @@ func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage)
 					"busid": file.Busid,
 					"url":   c.GetGroupFileUrl(m.GroupCode, file.Path, file.Busid),
 				},
-				"self_id": c.Uin,
-				"time":    time.Now().Unix(),
 			})
 			return
 		}
 	}
-	source := MessageSource{
-		SourceType: MessageSourceGroup,
-		PrimaryID:  uint64(m.GroupCode),
+	source := message.Source{
+		SourceType: message.SourceGroup,
+		PrimaryID:  m.GroupCode,
 	}
-	cqm := ToStringMessage(m.Elements, source, true)
+	cqm := toStringMessage(m.Elements, source)
 	id := bot.InsertGroupMessage(m)
 	log.Infof("收到群 %v(%v) 内 %v(%v) 的消息: %v (%v)", m.GroupName, m.GroupCode, m.Sender.DisplayName(), m.Sender.Uin, cqm, id)
 	gm := bot.formatGroupMessage(m)
 	if gm == nil {
 		return
 	}
-	gm["message_id"] = id
-	bot.dispatchEventMessage(gm)
+	gm.Others["message_id"] = id
+	bot.dispatch(gm)
 }
 
 func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEvent) {
 	m := e.Message
 	bot.checkMedia(m.Elements, m.Sender.Uin)
-	source := MessageSource{
-		SourceType: MessageSourcePrivate,
-		PrimaryID:  uint64(e.Session.Sender),
+	source := message.Source{
+		SourceType: message.SourcePrivate,
+		PrimaryID:  e.Session.Sender,
 	}
-	cqm := ToStringMessage(m.Elements, source, true)
+	cqm := toStringMessage(m.Elements, source)
 	bot.tempSessionCache.Store(m.Sender.Uin, e.Session)
 	id := m.Id
 	// todo(Mrs4s)
@@ -120,17 +148,12 @@ func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEven
 	// }
 	log.Infof("收到来自群 %v(%v) 内 %v(%v) 的临时会话消息: %v", m.GroupName, m.GroupCode, m.Sender.DisplayName(), m.Sender.Uin, cqm)
 	tm := global.MSG{
-		"post_type":    "message",
-		"message_type": "private",
-		"sub_type":     "group",
-		"temp_source":  e.Session.Source,
-		"message_id":   id,
-		"user_id":      m.Sender.Uin,
-		"message":      ToFormattedMessage(m.Elements, source, false),
-		"raw_message":  cqm,
-		"font":         0,
-		"self_id":      c.Uin,
-		"time":         time.Now().Unix(),
+		"temp_source": e.Session.Source,
+		"message_id":  id,
+		"user_id":     m.Sender.Uin,
+		"message":     ToFormattedMessage(m.Elements, source),
+		"raw_message": cqm,
+		"font":        0,
 		"sender": global.MSG{
 			"user_id":  m.Sender.Uin,
 			"group_id": m.GroupCode,
@@ -139,7 +162,7 @@ func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEven
 			"age":      0,
 		},
 	}
-	bot.dispatchEventMessage(tm)
+	bot.dispatchEvent("message/private/group", tm)
 }
 
 func (bot *CQBot) guildChannelMessageEvent(c *client.QQClient, m *message.GuildChannelMessage) {
@@ -149,31 +172,28 @@ func (bot *CQBot) guildChannelMessageEvent(c *client.QQClient, m *message.GuildC
 		return
 	}
 	channel := guild.FindChannel(m.ChannelId)
-	source := MessageSource{
-		SourceType: MessageSourceGuildChannel,
-		PrimaryID:  m.GuildId,
-		SubID:      m.ChannelId,
+	source := message.Source{
+		SourceType:  message.SourceGuildChannel,
+		PrimaryID:   int64(m.GuildId),
+		SecondaryID: int64(m.ChannelId),
 	}
-	log.Infof("收到来自频道 %v(%v) 子频道 %v(%v) 内 %v(%v) 的消息: %v", guild.GuildName, guild.GuildId, channel.ChannelName, m.ChannelId, m.Sender.Nickname, m.Sender.TinyId, ToStringMessage(m.Elements, source, true))
+	log.Infof("收到来自频道 %v(%v) 子频道 %v(%v) 内 %v(%v) 的消息: %v", guild.GuildName, guild.GuildId, channel.ChannelName, m.ChannelId, m.Sender.Nickname, m.Sender.TinyId, toStringMessage(m.Elements, source))
 	id := bot.InsertGuildChannelMessage(m)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "message",
-		"message_type": "guild",
-		"sub_type":     "channel",
+	ev := bot.event("message/guild/channel", global.MSG{
 		"guild_id":     fU64(m.GuildId),
 		"channel_id":   fU64(m.ChannelId),
 		"message_id":   id,
 		"user_id":      fU64(m.Sender.TinyId),
-		"message":      ToFormattedMessage(m.Elements, source, false), // todo: 增加对频道消息 Reply 的支持
-		"self_id":      bot.Client.Uin,
+		"message":      ToFormattedMessage(m.Elements, source), // todo: 增加对频道消息 Reply 的支持
 		"self_tiny_id": fU64(bot.Client.GuildService.TinyId),
-		"time":         m.Time,
 		"sender": global.MSG{
 			"user_id":  m.Sender.TinyId,
 			"tiny_id":  fU64(m.Sender.TinyId),
 			"nickname": m.Sender.Nickname,
 		},
 	})
+	ev.Time = m.Time
+	bot.dispatch(ev)
 }
 
 func (bot *CQBot) guildMessageReactionsUpdatedEvent(c *client.QQClient, e *client.GuildMessageReactionsUpdatedEvent) {
@@ -181,7 +201,7 @@ func (bot *CQBot) guildMessageReactionsUpdatedEvent(c *client.QQClient, e *clien
 	if guild == nil {
 		return
 	}
-	msgID := encodeGuildMessageID(e.GuildId, e.ChannelId, e.MessageId, MessageSourceGuildChannel)
+	msgID := encodeGuildMessageID(e.GuildId, e.ChannelId, e.MessageId, message.SourceGuildChannel)
 	str := fmt.Sprintf("频道 %v(%v) 消息 %v 表情贴片已更新: ", guild.GuildName, guild.GuildId, msgID)
 	currentReactions := make([]global.MSG, len(e.CurrentReactions))
 	for i, r := range e.CurrentReactions {
@@ -199,16 +219,12 @@ func (bot *CQBot) guildMessageReactionsUpdatedEvent(c *client.QQClient, e *clien
 		str += "无任何表情"
 	}
 	log.Infof(str)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":         "notice",
-		"notice_type":       "message_reactions_updated",
+	bot.dispatchEvent("notice/message_reactions_updated", global.MSG{
 		"guild_id":          fU64(e.GuildId),
 		"channel_id":        fU64(e.ChannelId),
 		"message_id":        msgID,
 		"operator_id":       fU64(e.OperatorId),
 		"current_reactions": currentReactions,
-		"time":              time.Now().Unix(),
-		"self_id":           bot.Client.Uin,
 		"self_tiny_id":      fU64(bot.Client.GuildService.TinyId),
 		"user_id":           e.OperatorId,
 	})
@@ -228,17 +244,13 @@ func (bot *CQBot) guildChannelMessageRecalledEvent(c *client.QQClient, e *client
 		log.Errorf("处理频道撤回事件时出现错误: 获取操作者资料时出现错误 %v", err)
 		return
 	}
-	msgID := encodeGuildMessageID(e.GuildId, e.ChannelId, e.MessageId, MessageSourceGuildChannel)
+	msgID := encodeGuildMessageID(e.GuildId, e.ChannelId, e.MessageId, message.SourceGuildChannel)
 	log.Infof("用户 %v(%v) 撤回了频道 %v(%v) 子频道 %v(%v) 的消息 %v", operator.Nickname, operator.TinyId, guild.GuildName, guild.GuildId, channel.ChannelName, channel.ChannelId, msgID)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "notice",
-		"notice_type":  "guild_channel_recall",
+	bot.dispatchEvent("notice/guild_channel_recall", global.MSG{
 		"guild_id":     fU64(e.GuildId),
 		"channel_id":   fU64(e.ChannelId),
 		"operator_id":  fU64(e.OperatorId),
 		"message_id":   msgID,
-		"time":         time.Now().Unix(),
-		"self_id":      bot.Client.Uin,
 		"self_tiny_id": fU64(bot.Client.GuildService.TinyId),
 		"user_id":      e.OperatorId,
 	})
@@ -250,14 +262,10 @@ func (bot *CQBot) guildChannelUpdatedEvent(c *client.QQClient, e *client.GuildCh
 		return
 	}
 	log.Infof("频道 %v(%v) 子频道 %v(%v) 信息已更新", guild.GuildName, guild.GuildId, e.NewChannelInfo.ChannelName, e.NewChannelInfo.ChannelId)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "notice",
-		"notice_type":  "channel_updated",
+	bot.dispatchEvent("notice/channel_updated", global.MSG{
 		"guild_id":     fU64(e.GuildId),
 		"channel_id":   fU64(e.ChannelId),
 		"operator_id":  fU64(e.OperatorId),
-		"time":         time.Now().Unix(),
-		"self_id":      bot.Client.Uin,
 		"self_tiny_id": fU64(bot.Client.GuildService.TinyId),
 		"user_id":      e.OperatorId,
 		"old_info":     convertChannelInfo(e.OldChannelInfo),
@@ -275,16 +283,12 @@ func (bot *CQBot) guildChannelCreatedEvent(c *client.QQClient, e *client.GuildCh
 		member = &client.GuildUserProfile{Nickname: "未知"}
 	}
 	log.Infof("频道 %v(%v) 内用户 %v(%v) 创建了子频道 %v(%v)", guild.GuildName, guild.GuildId, member.Nickname, member.TinyId, e.ChannelInfo.ChannelName, e.ChannelInfo.ChannelId)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "notice",
-		"notice_type":  "channel_created",
+	bot.dispatchEvent("notice/channel_created", global.MSG{
 		"guild_id":     fU64(e.GuildId),
 		"channel_id":   fU64(e.ChannelInfo.ChannelId),
 		"operator_id":  fU64(e.OperatorId),
-		"self_id":      bot.Client.Uin,
 		"self_tiny_id": fU64(bot.Client.GuildService.TinyId),
 		"user_id":      e.OperatorId,
-		"time":         time.Now().Unix(),
 		"channel_info": convertChannelInfo(e.ChannelInfo),
 	})
 }
@@ -299,16 +303,12 @@ func (bot *CQBot) guildChannelDestroyedEvent(c *client.QQClient, e *client.Guild
 		member = &client.GuildUserProfile{Nickname: "未知"}
 	}
 	log.Infof("频道 %v(%v) 内用户 %v(%v) 删除了子频道 %v(%v)", guild.GuildName, guild.GuildId, member.Nickname, member.TinyId, e.ChannelInfo.ChannelName, e.ChannelInfo.ChannelId)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "notice",
-		"notice_type":  "channel_destroyed",
+	bot.dispatchEvent("notice/channel_destroyed", global.MSG{
 		"guild_id":     fU64(e.GuildId),
 		"channel_id":   fU64(e.ChannelInfo.ChannelId),
 		"operator_id":  fU64(e.OperatorId),
-		"self_id":      bot.Client.Uin,
 		"self_tiny_id": fU64(bot.Client.GuildService.TinyId),
 		"user_id":      e.OperatorId,
-		"time":         time.Now().Unix(),
 		"channel_info": convertChannelInfo(e.ChannelInfo),
 	})
 }
@@ -332,22 +332,15 @@ func (bot *CQBot) groupMutedEvent(c *client.QQClient, e *client.GroupMuteEvent) 
 				formatGroupName(g), formatMemberName(g.FindMember(e.TargetUin)), formatMemberName(g.FindMember(e.OperatorUin)))
 		}
 	}
-
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
+	typ := "notice/group_ban/ban"
+	if e.Time == 0 {
+		typ = "notice/group_ban/lift_ban"
+	}
+	bot.dispatchEvent(typ, global.MSG{
 		"duration":    e.Time,
 		"group_id":    e.GroupCode,
-		"notice_type": "group_ban",
 		"operator_id": e.OperatorUin,
-		"self_id":     c.Uin,
 		"user_id":     e.TargetUin,
-		"time":        time.Now().Unix(),
-		"sub_type": func() string {
-			if e.Time == 0 {
-				return "lift_ban"
-			}
-			return "ban"
-		}(),
 	})
 }
 
@@ -356,16 +349,15 @@ func (bot *CQBot) groupRecallEvent(c *client.QQClient, e *client.GroupMessageRec
 	gid := db.ToGlobalID(e.GroupCode, e.MessageId)
 	log.Infof("群 %v 内 %v 撤回了 %v 的消息: %v.",
 		formatGroupName(g), formatMemberName(g.FindMember(e.OperatorUin)), formatMemberName(g.FindMember(e.AuthorUin)), gid)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
+
+	ev := bot.event("notice/group_recall", global.MSG{
 		"group_id":    e.GroupCode,
-		"notice_type": "group_recall",
-		"self_id":     c.Uin,
 		"user_id":     e.AuthorUin,
 		"operator_id": e.OperatorUin,
-		"time":        e.Time,
 		"message_id":  gid,
 	})
+	ev.Time = int64(e.Time)
+	bot.dispatch(ev)
 }
 
 func (bot *CQBot) groupNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
@@ -375,42 +367,27 @@ func (bot *CQBot) groupNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 		sender := group.FindMember(notify.Sender)
 		receiver := group.FindMember(notify.Receiver)
 		log.Infof("群 %v 内 %v 戳了戳 %v", formatGroupName(group), formatMemberName(sender), formatMemberName(receiver))
-		bot.dispatchEventMessage(global.MSG{
-			"post_type":   "notice",
-			"group_id":    group.Code,
-			"notice_type": "notify",
-			"sub_type":    "poke",
-			"self_id":     c.Uin,
-			"user_id":     notify.Sender,
-			"sender_id":   notify.Sender,
-			"target_id":   notify.Receiver,
-			"time":        time.Now().Unix(),
+		bot.dispatchEvent("notice/notify/poke", global.MSG{
+			"group_id":  group.Code,
+			"user_id":   notify.Sender,
+			"sender_id": notify.Sender,
+			"target_id": notify.Receiver,
 		})
 	case *client.GroupRedBagLuckyKingNotifyEvent:
 		sender := group.FindMember(notify.Sender)
 		luckyKing := group.FindMember(notify.LuckyKing)
 		log.Infof("群 %v 内 %v 的红包被抢完, %v 是运气王", formatGroupName(group), formatMemberName(sender), formatMemberName(luckyKing))
-		bot.dispatchEventMessage(global.MSG{
-			"post_type":   "notice",
-			"group_id":    group.Code,
-			"notice_type": "notify",
-			"sub_type":    "lucky_king",
-			"self_id":     c.Uin,
-			"user_id":     notify.Sender,
-			"sender_id":   notify.Sender,
-			"target_id":   notify.LuckyKing,
-			"time":        time.Now().Unix(),
+		bot.dispatchEvent("notice/notify/lucky_king", global.MSG{
+			"group_id":  group.Code,
+			"user_id":   notify.Sender,
+			"sender_id": notify.Sender,
+			"target_id": notify.LuckyKing,
 		})
 	case *client.MemberHonorChangedNotifyEvent:
 		log.Info(notify.Content())
-		bot.dispatchEventMessage(global.MSG{
-			"post_type":   "notice",
-			"group_id":    group.Code,
-			"notice_type": "notify",
-			"sub_type":    "honor",
-			"self_id":     c.Uin,
-			"user_id":     notify.Uin,
-			"time":        time.Now().Unix(),
+		bot.dispatchEvent("notice/notify/honor", global.MSG{
+			"group_id": group.Code,
+			"user_id":  notify.Uin,
 			"honor_type": func() string {
 				switch notify.Honor {
 				case client.Talkative:
@@ -439,15 +416,10 @@ func (bot *CQBot) friendNotifyEvent(c *client.QQClient, e client.INotifyEvent) {
 		} else {
 			log.Infof("好友 %v 戳了戳你.", friend.Nickname)
 		}
-		bot.dispatchEventMessage(global.MSG{
-			"post_type":   "notice",
-			"notice_type": "notify",
-			"sub_type":    "poke",
-			"self_id":     c.Uin,
-			"user_id":     notify.Sender,
-			"sender_id":   notify.Sender,
-			"target_id":   notify.Receiver,
-			"time":        time.Now().Unix(),
+		bot.dispatchEvent("notice/notify/poke", global.MSG{
+			"user_id":   notify.Sender,
+			"sender_id": notify.Sender,
+			"target_id": notify.Receiver,
 		})
 	}
 }
@@ -456,15 +428,10 @@ func (bot *CQBot) memberTitleUpdatedEvent(c *client.QQClient, e *client.MemberSp
 	group := c.FindGroup(e.GroupCode)
 	mem := group.FindMember(e.Uin)
 	log.Infof("群 %v(%v) 内成员 %v(%v) 获得了新的头衔: %v", group.Name, group.Code, mem.DisplayName(), mem.Uin, e.NewTitle)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "notify",
-		"sub_type":    "title",
-		"group_id":    group.Code,
-		"self_id":     c.Uin,
-		"user_id":     e.Uin,
-		"time":        time.Now().Unix(),
-		"title":       e.NewTitle,
+	bot.dispatchEvent("notice/notify/title", global.MSG{
+		"group_id": group.Code,
+		"user_id":  e.Uin,
+		"title":    e.NewTitle,
 	})
 }
 
@@ -476,14 +443,12 @@ func (bot *CQBot) friendRecallEvent(c *client.QQClient, e *client.FriendMessageR
 	} else {
 		log.Infof("好友 %v 撤回了消息: %v", e.FriendUin, gid)
 	}
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "friend_recall",
-		"self_id":     c.Uin,
-		"user_id":     e.FriendUin,
-		"time":        e.Time,
-		"message_id":  gid,
+	ev := bot.event("notice/friend_recall", global.MSG{
+		"user_id":    e.FriendUin,
+		"message_id": gid,
 	})
+	ev.Time = e.Time
+	bot.dispatch(ev)
 }
 
 func (bot *CQBot) offlineFileEvent(c *client.QQClient, e *client.OfflineFileEvent) {
@@ -492,23 +457,19 @@ func (bot *CQBot) offlineFileEvent(c *client.QQClient, e *client.OfflineFileEven
 		return
 	}
 	log.Infof("好友 %v(%v) 发送了离线文件 %v", f.Nickname, f.Uin, e.FileName)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "offline_file",
-		"user_id":     e.Sender,
+	bot.dispatchEvent("notice/offline_file", global.MSG{
+		"user_id": e.Sender,
 		"file": global.MSG{
 			"name": e.FileName,
 			"size": e.FileSize,
 			"url":  e.DownloadUrl,
 		},
-		"self_id": c.Uin,
-		"time":    time.Now().Unix(),
 	})
 }
 
 func (bot *CQBot) joinGroupEvent(c *client.QQClient, group *client.GroupInfo) {
 	log.Infof("Bot进入了群 %v.", formatGroupName(group))
-	bot.dispatchEventMessage(bot.groupIncrease(group.Code, 0, c.Uin))
+	bot.dispatch(bot.groupIncrease(group.Code, 0, c.Uin))
 }
 
 func (bot *CQBot) leaveGroupEvent(c *client.QQClient, e *client.GroupLeaveEvent) {
@@ -517,44 +478,33 @@ func (bot *CQBot) leaveGroupEvent(c *client.QQClient, e *client.GroupLeaveEvent)
 	} else {
 		log.Infof("Bot退出了群 %v.", formatGroupName(e.Group))
 	}
-	bot.dispatchEventMessage(bot.groupDecrease(e.Group.Code, c.Uin, e.Operator))
+	bot.dispatch(bot.groupDecrease(e.Group.Code, c.Uin, e.Operator))
 }
 
 func (bot *CQBot) memberPermissionChangedEvent(c *client.QQClient, e *client.MemberPermissionChangedEvent) {
-	st := func() string {
-		if e.NewPermission == client.Administrator {
-			return "set"
-		}
-		return "unset"
-	}()
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "group_admin",
-		"sub_type":    st,
-		"group_id":    e.Group.Code,
-		"user_id":     e.Member.Uin,
-		"time":        time.Now().Unix(),
-		"self_id":     c.Uin,
+	st := "unset"
+	if e.NewPermission == client.Administrator {
+		st = "set"
+	}
+	bot.dispatchEvent("notice/group_admin/"+st, global.MSG{
+		"group_id": e.Group.Code,
+		"user_id":  e.Member.Uin,
 	})
 }
 
 func (bot *CQBot) memberCardUpdatedEvent(c *client.QQClient, e *client.MemberCardUpdatedEvent) {
 	log.Infof("群 %v 的 %v 更新了名片 %v -> %v", formatGroupName(e.Group), formatMemberName(e.Member), e.OldCard, e.Member.CardName)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "group_card",
-		"group_id":    e.Group.Code,
-		"user_id":     e.Member.Uin,
-		"card_new":    e.Member.CardName,
-		"card_old":    e.OldCard,
-		"time":        time.Now().Unix(),
-		"self_id":     c.Uin,
+	bot.dispatchEvent("notice/group_card", global.MSG{
+		"group_id": e.Group.Code,
+		"user_id":  e.Member.Uin,
+		"card_new": e.Member.CardName,
+		"card_old": e.OldCard,
 	})
 }
 
 func (bot *CQBot) memberJoinEvent(_ *client.QQClient, e *client.MemberJoinGroupEvent) {
 	log.Infof("新成员 %v 进入了群 %v.", formatMemberName(e.Member), formatGroupName(e.Group))
-	bot.dispatchEventMessage(bot.groupIncrease(e.Group.Code, 0, e.Member.Uin))
+	bot.dispatch(bot.groupIncrease(e.Group.Code, 0, e.Member.Uin))
 }
 
 func (bot *CQBot) memberLeaveEvent(_ *client.QQClient, e *client.MemberLeaveGroupEvent) {
@@ -563,65 +513,47 @@ func (bot *CQBot) memberLeaveEvent(_ *client.QQClient, e *client.MemberLeaveGrou
 	} else {
 		log.Infof("成员 %v 离开了群 %v.", formatMemberName(e.Member), formatGroupName(e.Group))
 	}
-	bot.dispatchEventMessage(bot.groupDecrease(e.Group.Code, e.Member.Uin, e.Operator))
+	bot.dispatch(bot.groupDecrease(e.Group.Code, e.Member.Uin, e.Operator))
 }
 
 func (bot *CQBot) friendRequestEvent(c *client.QQClient, e *client.NewFriendRequest) {
 	log.Infof("收到来自 %v(%v) 的好友请求: %v", e.RequesterNick, e.RequesterUin, e.Message)
 	flag := strconv.FormatInt(e.RequestId, 10)
 	bot.friendReqCache.Store(flag, e)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "request",
-		"request_type": "friend",
-		"user_id":      e.RequesterUin,
-		"comment":      e.Message,
-		"flag":         flag,
-		"time":         time.Now().Unix(),
-		"self_id":      c.Uin,
+	bot.dispatchEvent("request/friend", global.MSG{
+		"user_id": e.RequesterUin,
+		"comment": e.Message,
+		"flag":    flag,
 	})
 }
 
 func (bot *CQBot) friendAddedEvent(c *client.QQClient, e *client.NewFriendEvent) {
 	log.Infof("添加了新好友: %v(%v)", e.Friend.Nickname, e.Friend.Uin)
 	bot.tempSessionCache.Delete(e.Friend.Uin)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "friend_add",
-		"self_id":     c.Uin,
-		"user_id":     e.Friend.Uin,
-		"time":        time.Now().Unix(),
+	bot.dispatchEvent("notice/friend_add", global.MSG{
+		"user_id": e.Friend.Uin,
 	})
 }
 
 func (bot *CQBot) groupInvitedEvent(c *client.QQClient, e *client.GroupInvitedRequest) {
 	log.Infof("收到来自群 %v(%v) 内用户 %v(%v) 的加群邀请.", e.GroupName, e.GroupCode, e.InvitorNick, e.InvitorUin)
 	flag := strconv.FormatInt(e.RequestId, 10)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "request",
-		"request_type": "group",
-		"sub_type":     "invite",
-		"group_id":     e.GroupCode,
-		"user_id":      e.InvitorUin,
-		"comment":      "",
-		"flag":         flag,
-		"time":         time.Now().Unix(),
-		"self_id":      c.Uin,
+	bot.dispatchEvent("request/group/invite", global.MSG{
+		"group_id": e.GroupCode,
+		"user_id":  e.InvitorUin,
+		"comment":  "",
+		"flag":     flag,
 	})
 }
 
 func (bot *CQBot) groupJoinReqEvent(c *client.QQClient, e *client.UserJoinGroupRequest) {
 	log.Infof("群 %v(%v) 收到来自用户 %v(%v) 的加群请求.", e.GroupName, e.GroupCode, e.RequesterNick, e.RequesterUin)
 	flag := strconv.FormatInt(e.RequestId, 10)
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":    "request",
-		"request_type": "group",
-		"sub_type":     "add",
-		"group_id":     e.GroupCode,
-		"user_id":      e.RequesterUin,
-		"comment":      e.Message,
-		"flag":         flag,
-		"time":         time.Now().Unix(),
-		"self_id":      c.Uin,
+	bot.dispatchEvent("request/group/add", global.MSG{
+		"group_id": e.GroupCode,
+		"user_id":  e.RequesterUin,
+		"comment":  e.Message,
+		"flag":     flag,
 	})
 }
 
@@ -631,17 +563,13 @@ func (bot *CQBot) otherClientStatusChangedEvent(c *client.QQClient, e *client.Ot
 	} else {
 		log.Infof("Bot 账号在客户端 %v (%v) 登出.", e.Client.DeviceName, e.Client.DeviceKind)
 	}
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
-		"notice_type": "client_status",
-		"online":      e.Online,
+	bot.dispatchEvent("notice/client_status", global.MSG{
+		"online": e.Online,
 		"client": global.MSG{
 			"app_id":      e.Client.AppId,
 			"device_name": e.Client.DeviceName,
 			"device_kind": e.Client.DeviceKind,
 		},
-		"self_id": c.Uin,
-		"time":    time.Now().Unix(),
 	})
 }
 
@@ -668,61 +596,44 @@ func (bot *CQBot) groupEssenceMsg(c *client.QQClient, e *client.GroupDigestEvent
 	if e.OperatorUin == bot.Client.Uin {
 		return
 	}
-	bot.dispatchEventMessage(global.MSG{
-		"post_type":   "notice",
+	subtype := "delete"
+	if e.OperationType == 1 {
+		subtype = "add"
+	}
+	bot.dispatchEvent("notice/essence/"+subtype, global.MSG{
 		"group_id":    e.GroupCode,
-		"notice_type": "essence",
-		"sub_type": func() string {
-			if e.OperationType == 1 {
-				return "add"
-			}
-			return "delete"
-		}(),
-		"self_id":     c.Uin,
 		"sender_id":   e.SenderUin,
 		"operator_id": e.OperatorUin,
-		"time":        time.Now().Unix(),
 		"message_id":  gid,
 	})
 }
 
-func (bot *CQBot) groupIncrease(groupCode, operatorUin, userUin int64) global.MSG {
-	return global.MSG{
-		"post_type":   "notice",
-		"notice_type": "group_increase",
+func (bot *CQBot) groupIncrease(groupCode, operatorUin, userUin int64) *event {
+	return bot.event("notice/group_increase/approve", global.MSG{
 		"group_id":    groupCode,
 		"operator_id": operatorUin,
-		"self_id":     bot.Client.Uin,
-		"sub_type":    "approve",
-		"time":        time.Now().Unix(),
 		"user_id":     userUin,
-	}
+	})
 }
 
-func (bot *CQBot) groupDecrease(groupCode, userUin int64, operator *client.GroupMemberInfo) global.MSG {
-	return global.MSG{
-		"post_type":   "notice",
-		"notice_type": "group_decrease",
-		"group_id":    groupCode,
-		"operator_id": func() int64 {
-			if operator != nil {
-				return operator.Uin
-			}
-			return userUin
-		}(),
-		"self_id": bot.Client.Uin,
-		"sub_type": func() string {
-			if operator != nil {
-				if userUin == bot.Client.Uin {
-					return "kick_me"
-				}
-				return "kick"
-			}
-			return "leave"
-		}(),
-		"time":    time.Now().Unix(),
-		"user_id": userUin,
+func (bot *CQBot) groupDecrease(groupCode, userUin int64, operator *client.GroupMemberInfo) *event {
+	op := userUin
+	if operator != nil {
+		op = operator.Uin
 	}
+	subtype := "leave"
+	if operator != nil {
+		if userUin == bot.Client.Uin {
+			subtype = "kick_me"
+		} else {
+			subtype = "kick"
+		}
+	}
+	return bot.event("notice/group_decrease/"+subtype, global.MSG{
+		"group_id":    groupCode,
+		"operator_id": op,
+		"user_id":     userUin,
+	})
 }
 
 func (bot *CQBot) checkMedia(e []message.IMessageElement, sourceID int64) {
