@@ -79,7 +79,7 @@ func DownloadFile(url, path string, limit int64, headers map[string]string) erro
 	if limit > 0 && resp.ContentLength > limit {
 		return ErrOverSize
 	}
-	_, err = io.Copy(file, resp.Body)
+	_, err = file.ReadFrom(resp.Body)
 	if err != nil {
 		return err
 	}
@@ -107,7 +107,7 @@ func DownloadFileMultiThreading(url, path string, limit int64, threadCount int, 
 				return err
 			}
 			defer file.Close()
-			if _, err = io.Copy(file, s); err != nil {
+			if _, err = file.ReadFrom(s); err != nil {
 				return err
 			}
 			return errUnsupportedMultiThreading
