@@ -7,13 +7,12 @@ import (
 	"fmt"
 	"hash"
 	"io"
+	"math"
 	"os"
 	"path/filepath"
 	"runtime"
 	"strings"
 
-	"github.com/dustin/go-humanize"
-	"github.com/kardianos/osext"
 	"github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
 
@@ -147,13 +146,33 @@ func (wc *writeSumCounter) Write(p []byte) (int, error) {
 	wc.total += uint64(n)
 	wc.hash.Write(p)
 	fmt.Printf("\r                                    ")
-	fmt.Printf("\rDownloading... %s complete", humanize.Bytes(wc.total))
+	fmt.Printf("\rDownloading... %s complete", humanBytes(wc.total))
 	return n, nil
+}
+
+func logn(n, b float64) float64 {
+	return math.Log(n) / math.Log(b)
+}
+
+func humanBytes(s uint64) string {
+	sizes := []string{"B", "kB", "MB", "GB"} // GB对于go-cqhttp来说已经够用了
+	if s < 10 {
+		return fmt.Sprintf("%d B", s)
+	}
+	e := math.Floor(logn(float64(s), 1000))
+	suffix := sizes[int(e)]
+	val := math.Floor(float64(s)/math.Pow(1000, e)*10+0.5) / 10
+	f := "%.0f %s"
+	if val < 10 {
+		f = "%.1f %s"
+	}
+	return fmt.Sprintf(f, val, suffix)
 }
 
 // FromStream copy form getlantern/go-update
 func fromStream(updateWith io.Reader) (err error, errRecover error) {
-	updatePath, err := osext.Executable()
+	updatePath, err := os.Executable()
+	updatePath = filepath.Clean(updatePath)
 	if err != nil {
 		return
 	}
@@ -169,7 +188,7 @@ func fromStream(updateWith io.Reader) (err error, errRecover error) {
 	}
 	// We won't log this error, because it's always going to happen.
 	defer func() { _ = fp.Close() }()
-	if _, err = io.Copy(fp, bufio.NewReader(updateWith)); err != nil {
+	if _, err = bufio.NewReader(updateWith).WriteTo(fp); err != nil {
 		logrus.Errorf("Unable to copy data: %v\n", err)
 	}
 
