@@ -6,12 +6,11 @@ package selfupdate
 import (
 	"archive/tar"
 	"bytes"
+	"compress/gzip"
 	"crypto/sha256"
 	"errors"
 	"io"
 	"net/http"
-
-	"github.com/klauspost/compress/gzip"
 )
 
 // update go-cqhttp自我更新
