@@ -21,12 +21,12 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case ".ocr_image", "ocr_image":
 		p0 := p.Get("image").String()
 		return c.bot.CQOcrImage(p0)
+	case "_get_group_notice":
+		p0 := p.Get("group_id").Int()
+		return c.bot.CQGetGroupMemo(p0)
 	case "_get_model_show":
 		p0 := p.Get("model").String()
 		return c.bot.CQGetModelShow(p0)
-	case "_get_vip_info":
-		p0 := p.Get("user_id").Int()
-		return c.bot.CQGetVipInfo(p0)
 	case "_send_group_notice":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("content").String()
@@ -195,6 +195,12 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case "reload_event_filter":
 		p0 := p.Get("file").String()
 		return c.bot.CQReloadEventFilter(p0)
+	case "send_forward_msg":
+		p0 := p.Get("group_id").Int()
+		p1 := p.Get("user_id").Int()
+		p2 := p.Get("messages")
+		p3 := p.Get("message_type").String()
+		return c.bot.CQSendForwardMessage(p0, p1, p2, p3)
 	case "send_group_forward_msg":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("messages")
@@ -204,6 +210,9 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p1 := p.Get("message")
 		p2 := p.Get("auto_escape").Bool()
 		return c.bot.CQSendGroupMessage(p0, p1, p2)
+	case "send_group_sign":
+		p0 := p.Get("group_id").Int()
+		return c.bot.CQSendGroupSign(p0)
 	case "send_guild_channel_msg":
 		p0 := p.Get("guild_id").Uint()
 		p1 := p.Get("channel_id").Uint()
@@ -217,6 +226,10 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p3 := p.Get("message_type").String()
 		p4 := p.Get("auto_escape").Bool()
 		return c.bot.CQSendMessage(p0, p1, p2, p3, p4)
+	case "send_private_forward_msg":
+		p0 := p.Get("user_id").Int()
+		p1 := p.Get("messages")
+		return c.bot.CQSendPrivateForwardMessage(p0, p1)
 	case "send_private_msg":
 		p0 := p.Get("user_id").Int()
 		p1 := p.Get("group_id").Int()
@@ -304,6 +317,13 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p2 := p.Get("role_id").Uint()
 		p3 := p.Get("users")
 		return c.bot.CQSetGuildMemberRole(p0, p1, p2, p3)
+	case "set_qq_profile":
+		p0 := p.Get("nickname")
+		p1 := p.Get("company")
+		p2 := p.Get("email")
+		p3 := p.Get("college")
+		p4 := p.Get("personal_note")
+		return c.bot.CQSetQQProfile(p0, p1, p2, p3, p4)
 	case "update_guild_role":
 		p0 := p.Get("guild_id").Uint()
 		p1 := p.Get("role_id").Uint()
